Moves bus exceptions. Closes #1334
