+++ app/Bus/Exceptions/Subscriber/AlreadySubscribedException.php
@@ -9,14 +9,16 @@
  * file that was distributed with this source code.
  */
 
-namespace CachetHQ\Cachet\Exceptions;
+namespace CachetHQ\Cachet\Bus\Exceptions\Subscriber;
 
+use CachetHQ\Cachet\Bus\Exceptions\ExceptionInterface;
 use Exception;
 
 /**
  * This is the already subscribed exception class.
  *
  * @author Graham Campbell <graham@alt-three.com>
+ * @author James Brooks <james@alt-three.com>
  */
 class AlreadySubscribedException extends Exception implements ExceptionInterface
 {
